+++ src/nice_int/nice_u32.rs
@@ -102,30 +102,37 @@ impl NiceU32 {
 		out
 	}
 
+	#[allow(clippy::cast_possible_truncation)] // One digit always fits u8.
 	/// # Parse.
 	///
 	/// This handles the actual crunching.
 	fn parse(&mut self, mut num: u32) {
 		let ptr = self.inner.as_mut_ptr();
 
-		while num >= 1000 {
+		while 999 < num {
 			let (div, rem) = crate::div_mod_u32(num, 1000);
-			unsafe { super::write_u8_3(ptr.add(self.from - 3), rem as usize); }
-			num = div;
 			self.from -= 4;
+			unsafe { super::write_u8_3(ptr.add(self.from + 1), rem as u16); }
+			num = div;
 		}
 
-		if num >= 100 {
+		if 99 < num {
 			self.from -= 3;
-			unsafe { super::write_u8_3(ptr.add(self.from), num as usize); }
+			unsafe { super::write_u8_3(ptr.add(self.from), num as u16); }
 		}
-		else if num >= 10 {
+		else if 9 < num {
 			self.from -= 2;
-			unsafe { super::write_u8_2(ptr.add(self.from), num as usize); }
+			unsafe {
+				std::ptr::copy_nonoverlapping(
+					crate::double(num as usize),
+					ptr.add(self.from),
+					2
+				);
+			}
 		}
 		else {
 			self.from -= 1;
-			unsafe { super::write_u8_1(ptr.add(self.from), num as usize); }
+			unsafe { std::ptr::write(ptr.add(self.from), num as u8 + b'0'); }
 		}
 	}
 }